# projectT4T/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/GetFieldFacilitator": {
            "get": {
                "tags": [
                    "FieldFacilitator"
                ],
                "summary": "Get Field Facilitator",
                "operationId": "GetFieldFacilitator",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "type": "integer"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/GetFieldFacilitatorDetail": {
            "get": {
                "tags": [
                    "FieldFacilitator"
                ],
                "summary": "Get Field Facilitator Detail",
                "operationId": "GetFieldFacilitatorDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/AddFieldFacilitator": {
            "post": {
                "tags": [
                    "FieldFacilitator"
                ],
                "summary": "Add Field Facilitator",
                "operationId": "AddFieldFacilitator",
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Add Field Facilitator",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Mangga"
                                },
                                "birthday": {
                                    "type": "date",
                                    "example": "1990-01-30"
                                },
                                "religion": {
                                    "type": "string",
                                    "example": "islam"
                                },
                                "gender": {
                                    "type": "string",
                                    "example": "male/female"
                                },
                                "ktp_no": {
                                    "type": "string",
                                    "example": "33101700020001"
                                },
                                "address": {
                                    "type": "string",
                                    "example": "Jl Cemara 11"
                                },
                                "village": {
                                    "type": "string",
                                    "example": "32.04.30.01"
                                },
                                "kecamatan": {
                                    "type": "string",
                                    "example": "32.04.30.01"
                                },
                                "city": {
                                    "type": "string",
                                    "example": "32.04"
                                },
                                "province": {
                                    "type": "string",
                                    "example": "JT"
                                },
                                "mu_no": {
                                    "type": "string",
                                    "example": "023"
                                },
                                "target_area": {
                                    "type": "string",
                                    "example": "023001"
                                },
                                "active": {
                                    "type": "string",
                                    "example": "1"
                                },
                                "user_id": {
                                    "type": "string",
                                    "example": "023"
                                },
                                "marrital": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "join_date": {
                                    "type": "date",
                                    "example": "Nullable"
                                },
                                "phone": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "post_code": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "bank_account": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "bank_branch": {
                                    "type": "date",
                                    "example": "Nullable"
                                },
                                "bank_name": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "ff_photo": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "ff_photo_path": {
                                    "type": "string",
                                    "example": "Nullable"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/UpdateFieldFacilitator": {
            "post": {
                "tags": [
                    "FieldFacilitator"
                ],
                "summary": "Update Field Facilitator",
                "operationId": "UpdateFieldFacilitator",
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Update Field Facilitator",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Mangga"
                                },
                                "birthday": {
                                    "type": "date",
                                    "example": "1990-01-30"
                                },
                                "religion": {
                                    "type": "string",
                                    "example": "islam"
                                },
                                "gender": {
                                    "type": "string",
                                    "example": "male/female"
                                },
                                "ktp_no": {
                                    "type": "string",
                                    "example": "33101700020001"
                                },
                                "address": {
                                    "type": "string",
                                    "example": "Jl Cemara 11"
                                },
                                "village": {
                                    "type": "string",
                                    "example": "32.04.30.01"
                                },
                                "kecamatan": {
                                    "type": "string",
                                    "example": "32.04.30.01"
                                },
                                "city": {
                                    "type": "string",
                                    "example": "32.04"
                                },
                                "province": {
                                    "type": "string",
                                    "example": "JT"
                                },
                                "mu_no": {
                                    "type": "string",
                                    "example": "023"
                                },
                                "target_area": {
                                    "type": "string",
                                    "example": "023001"
                                },
                                "active": {
                                    "type": "string",
                                    "example": "1"
                                },
                                "user_id": {
                                    "type": "string",
                                    "example": "023"
                                },
                                "marrital": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "join_date": {
                                    "type": "date",
                                    "example": "Nullable"
                                },
                                "phone": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "post_code": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "bank_account": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "bank_branch": {
                                    "type": "date",
                                    "example": "Nullable"
                                },
                                "bank_name": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "ff_photo": {
                                    "type": "string",
                                    "example": "Nullable"
                                },
                                "ff_photo_path": {
                                    "type": "string",
                                    "example": "Nullable"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/DeleteFieldFacilitator": {
            "post": {
                "tags": [
                    "FieldFacilitator"
                ],
                "summary": "Delete Field Facilitator",
                "operationId": "DeleteFieldFacilitator",
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Delete Field Facilitator",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "apiAuth": {
            "type": "apiKey",
            "description": "Login with email and password to get the authentication token",
            "name": "Authorization",
            "in": "header"
        }
    }
}